+++ test/native-suite-count
@@ -1,1 +1,1 @@
-30
+32
